Created project, prepareMeshFromHeightMap Method
prepareMeshFromHeightMap method so far uses a simple algorithm to create vertices from float[,] heightmap so that x and z coordinates of Vector3s inserted into the array correspond the coordinates of the heightmap and the y coordinate of the Vector3s corresponds to the value of the heightmap.

The triangles are created by starting from zero, taking the current point, point to the right, point below for first triangle. then increasing the current point and taking the new current point, the point below it and the point to the lower left.  Then if this is the rightmost point, the current is increased to take the left most point again.

Then the vertices and triangles are put into a mesh.

I plan to create unit tests for this method.

The newHeightMap methos is a placeholder.